Combine SettingSetTime with SettingSetDate
