+++ src/displayapp/Apps.h
@@ -34,8 +34,7 @@ namespace Pinetime {
       SettingDisplay,
       SettingWakeUp,
       SettingSteps,
-      SettingSetDate,
-      SettingSetTime,
+      SettingSetDateTime,
       SettingChimes,
       SettingShakeThreshold,
       SettingBluetooth,
